move swipeable into multiple files
