+++ packages/react-native-gesture-handler/src/components/ReanimatedSwipeable/index.ts
@@ -1,0 +1,6 @@
+export {
+  type SwipeableProps,
+  type SwipeableMethods,
+  SwipeDirection,
+} from './ReanimatedSwipeableProps';
+export { default } from './ReanimatedSwipeable';
